Massive refactoring of the workDjango process
This commit removes all the submodules and replaces the currently
developed modules with symlinks.  The idea is that to develop the
modules they will be cloned outside this repository and the links will
point to the module code.  This also allows the modules to be eggs but
are developed as though they were installed.